Fix typo on stack so top is an integer type
Necessary so top can be a negative value when the
stack is empty.

+++ 27-stack_array/stack.h
@@ -14,7 +14,7 @@
  * size: is the size of the
  */
 struct stack {
-    size_t top;
+    long long int top;
     size_t size;
     char *stack_data;
 };
